requirements.txt
dj-database-url==0.5.0
Django==2.0.6
django-filter==1.1.0
django-heroku==0.3.1
gunicorn==19.9.0
psycopg2==2.7.5
psycopg2-binary==2.7.5
python-dateutil==2.7.3
pytz==2018.5
six==1.11.0
whitenoise==3.3.1
